python-fd: merge postgres plugin into one file
