+++ debian/bareos-filedaemon-postgresql-python-plugin.install.in
@@ -1,3 +1,2 @@
 @plugindir@/bareos-fd-postgres.py*
-@plugindir@/BareosFdPluginPostgres.py*
 @plugindir@/bareos-fd-postgresql.py*
